backend/requirements.txt
# HubSpot AI Wrapper - Backend (FastAPI)
# Python 3.11+ recommended

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client for HubSpot & external APIs
httpx>=0.26.0

# Supabase
supabase>=2.4.0

# Environment
python-dotenv==1.0.1

# JWT
python-jose[cryptography]==3.3.0

# LLM (OpenAI; for Anthropic use anthropic SDK instead)
openai==1.12.0

# Optional: Anthropic (uncomment if using Claude)
# anthropic==0.18.1
